feat(home-nav): Home 拆分为 Studio + Dashboard 双子页面并迁移到嵌套段路由;
路由结构变更（Plan §3 方案 A）：
- 根 / → server-side redirect('/studio')，无 hydration mismatch；
- app/(home)/layout.tsx 承载共享 Suspense + HomeNav；
- app/(home)/studio/page.tsx 平移自原 app/page.tsx（聊天界面零行为变化，
  sessionId / CopilotKit Provider / NdjsonHttpAgent 完整保留）；
- app/(home)/dashboard/ 路由占位，下个 commit 填充内容。

HomeNav 仿 MemoryNav 风格的 rounded-full 胶囊 Tab。设计取舍：sessionId 仅
Studio 携带，Dashboard URL 干净；在 Tab 之间切换时若有 sessionId 自动保留在
Studio 链接上，避免会话状态丢失。
